fix: silence lab-fs clippy failures under -D warnings
CI Clippy on main failed on unused locals, too_many_arguments,
identity ops, and type complexity in the real FS parsers.

+++ crates/lab-fs/src/ntfs/runlist.rs
@@ -27,7 +27,7 @@ pub fn decode_runlist(data: &[u8]) -> LabResult<Vec<DataRun>> {
                 detail: "truncated runlist".into(),
             });
         }
-        let vcn_length = read_le_int(&data[i..i + len_size]) as u64;
+        let vcn_length = read_le_int(&data[i..i + len_size]);
         i += len_size;
         let lcn = if off_size == 0 {
             None
